gwssrv: fix bug on refresh rectandle routine.
libio01: Small changes on types.
Small changes on Makefiles for some libs.

+++ gramci/lib/libio01/include/libio.h
@@ -5,13 +5,13 @@
  */
 
 
-//retorna o valor.
-unsigned char gde_inport8 (unsigned short port);
-unsigned short gde_inport16 (unsigned short port);
-unsigned long gde_inport32 (unsigned short port);
+// Retorna o valor.
+unsigned char  gde_inport8  (unsigned int port);
+unsigned short gde_inport16 (unsigned int port);
+unsigned long  gde_inport32 (unsigned int port);
 
-void gde_outport8 ( unsigned short port, unsigned char value);
-void gde_outport16 ( unsigned short port, unsigned short value);
-void gde_outport32 ( unsigned short port, unsigned long value);
+void gde_outport8  ( unsigned int port, unsigned char  value);
+void gde_outport16 ( unsigned int port, unsigned short value);
+void gde_outport32 ( unsigned int port, unsigned long  value);
